Cleaning up variables
